fix: item 아이디 타입 변경

+++ backend/src/main/java/io/watssuggang/voda/pet/dto/res/ItemResponse.java
@@ -7,7 +7,7 @@ import lombok.RequiredArgsConstructor;
 @Getter
 public class ItemResponse {
 
-    private final Long itemId;
+    private final Integer itemId;
     private final String imgURl;
     private final String name;
 }
